refactor: isolate security models from entities
