+++ src/main/java/com/academy/e_commerce/config/security/UserPrincipal.java
@@ -1,5 +1,6 @@
-package com.academy.e_commerce.model;
+package com.academy.e_commerce.config.security;
 
+import com.academy.e_commerce.model.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
